+++ src/serviceWorker.ts
@@ -14,7 +14,7 @@
 // "./service-worker.js" file changes). Therefore we use a version number here,
 // so we can force updates.
 //
-const CATE_PWA_VERSION = "3.0.0-dev.7";
+const CATE_PWA_VERSION = "3.0.0";
 
 console.debug(`Cate PWA version ${CATE_PWA_VERSION}`);